Add files via upload
Mini-project: Perceptron 
The training set in perceptron.data contains the Iris dataset limited to Iris-setosa and Iris-versicolor. perceptron.test.data contains the test set. Implement the perceptron and train it to classify the two species. Test with the test set and output the accuracy. The program should have the following capabilities: ˆ Loading any dataset in csv format, where the last column is the class. The number of weights should be adjusted to the dataset. (!!) ˆ Picking the learning rate. ˆ Simple UI to manually input vectors to classify.